refactor: update import paths to use the 'components' directory for better organization
