+++ src/components/pages/StreamManagerPage.tsx
@@ -1,7 +1,7 @@
-import AdManager from '@/molecules/AdManager';
-import StreamInfoEditor from '@/molecules/StreamInfoEditor';
-import Timers from '@/molecules/Timers';
-import UpcomingStream from '@/molecules/UpcomingStream';
+import AdManager from '@/components/molecules/AdManager';
+import StreamInfoEditor from '@/components/molecules/StreamInfoEditor';
+import Timers from '@/components/molecules/Timers';
+import UpcomingStream from '@/components/molecules/UpcomingStream';
 import findNextStream from '@/scheduling/findNextStream';
 import useProfile from '@/useProfile';
 import Alert from '@mui/material/Alert';
